fixed package names vs physical directory + organize imports
