+++ src/com/github/randerzander/StormCommon/bolts/PhoenixLookupBolt.java
@@ -1,33 +1,20 @@
 package com.github.randerzander.StormCommon.bolts;
 
-import com.github.randerzander.StormCommon.Utils;
-
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import java.sql.DatabaseMetaData;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.List;
 import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 import java.util.Properties;
-import java.util.Arrays;
-import java.io.File;
-
-import org.apache.commons.lang3.StringUtils;
 
 import backtype.storm.task.OutputCollector;
 import backtype.storm.task.TopologyContext;
 import backtype.storm.topology.IRichBolt;
 import backtype.storm.topology.OutputFieldsDeclarer;
-import backtype.storm.tuple.Fields;
 import backtype.storm.tuple.Tuple;
-import backtype.storm.Config;
-
-import org.apache.phoenix.jdbc.PhoenixDriver;
 
 public class PhoenixLookupBolt implements IRichBolt {
   private String jdbcURL;
