+++ src/com/github/randerzander/StormCommon/bolts/RollingWindowBolt.java
@@ -1,34 +1,25 @@
 package com.github.randerzander.StormCommon.bolts;
 
-import com.github.randerzander.StormCommon.Utils;
-
 import java.sql.Connection;
 import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import java.sql.DatabaseMetaData;
 import java.sql.ResultSet;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.ArrayList;
 import java.util.Iterator;
+import java.util.Map;
 import java.util.Properties;
-import java.util.Collections;
-import org.apache.commons.lang3.StringUtils;
-import org.joda.time.format.DateTimeFormatter;
-import org.joda.time.format.DateTimeFormat;
+
 import org.joda.time.DateTime;
+import org.joda.time.format.DateTimeFormat;
+import org.joda.time.format.DateTimeFormatter;
 
 import backtype.storm.task.OutputCollector;
 import backtype.storm.task.TopologyContext;
 import backtype.storm.topology.IRichBolt;
 import backtype.storm.topology.OutputFieldsDeclarer;
-import backtype.storm.tuple.Tuple;
 import backtype.storm.tuple.Fields;
-import backtype.storm.tuple.Values;
-import backtype.storm.Config;
+import backtype.storm.tuple.Tuple;
 
-import org.apache.phoenix.jdbc.PhoenixDriver;
+import com.github.randerzander.StormCommon.utils.Utils;
 
 public class RollingWindowBolt implements IRichBolt {
   private String jdbcURL;
